backend/requirements-dev.txt
asgiref==3.8.1
attrs==24.3.0
autopep8==2.3.1
black==24.10.0
boto3==1.28.7
botocore==1.31.85
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.8
cryptography==44.0.0
defusedxml==0.8.0rc2
diff-match-patch==20241021
distlib==0.3.9
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.2.16
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.13.2
django-appconf==1.0.6
django-avatar==8.0.1
django-browser-reload==1.12.0
django-cors-headers==3.13.0
django-crum==0.7.9
django-debug-toolbar==4.3.0
django-editorjs==0.2.1
django-filter==23.5
django-import-export==4.1.1
django-modeladmin-reorder==0.3.1
django-object-actions==4.3.0
django-quill-editor==0.1.42
django-richtextfield==1.6.1
django-storages==1.13.2
django-tinymce==4.1.0
django-unfold==0.39.0
django-widget-tweaks==1.5.0
django_app_novadata==0.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-quill-editor==0.1.42
djoser==2.3.1
dnspython==2.7.0
docutils==0.21.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
filelock==3.16.1
flake8==7.1.1
identify==2.6.3
idna==3.10
inflection==0.5.1
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.5.0
Markdown==3.7
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
model-bakery==1.20.0
more-itertools==10.5.0
mypy==1.14.0
mypy-extensions==1.0.0
nh3==0.2.20
nodeenv==1.9.1
novadata_utils==0.3.9
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
pkginfo==1.12.0
platformdirs==4.3.6
pre_commit==4.0.1
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0.2
readme_renderer==44.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.6.2
SecretStorage==3.3.3
setuptools==75.6.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.4.2
static3==0.7.0
tablib==3.5.0
twine==6.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.0.7
virtualenv==20.28.0
wheel==0.45.1
